requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.6.6
decorator==5.1.1
Django==4.1.5
django-crispy-forms==1.14.0
django-filter==23.1
django-tinymce==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-yasg==1.21.5
executing==1.2.0
idna==3.4
inflection==0.5.1
ipykernel==6.21.3
ipython==8.11.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pythonanywhere==0.11.0
pytz==2022.7.1
pywin32==305
pyzmq==25.0.1
requests==2.28.2
ruamel.yaml==0.17.21
schema==0.7.5
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tabulate==0.9.0
tornado==6.2
traitlets==5.9.0
typer==0.7.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
